Finishes reformating view routes.
