+++ app/views/account/__init__.py
@@ -1,1 +1,2 @@
 from . import auth
+from . import profile
