+++ app/views/register/__init__.py
@@ -1,1 +1,3 @@
 from . import register
+# Route disabled
+# from . import preregister
