+++ app/views/team/__init__.py
@@ -1,0 +1,4 @@
+from . import team
+from . import rename
+from . import join
+from . import leave
